requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
web3==6.20.0
eth-account>=0.8.0,<0.13
pydantic[email]==2.8.2
email-validator
python-dotenv==1.0.1
cryptography==42.0.8
httpx==0.27.0
python-multipart==0.0.9
visidata
pyttsx3==2.90
python-jose[cryptography]==3.3.0
passlib[bcrypt]
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
slowapi==0.1.9
boto3==1.34.0
PyMuPDF==1.24.7
pdfplumber==0.11.0
pydub==0.25.1
moviepy==1.0.3
redis==5.0.8
aiofiles==24.1.0
pydantic-settings==2.4.0
asyncpg==0.29.0
pytest==8.3.2
pytest-asyncio==0.23.8
httpx==0.27.0
structlog
aiosqlite==0.19.0
# Added for DALS integration
